Refactor endpoint trace graph condition helper
